Improving recent login detection algorithm
The algorithm now treat the case of old and new lists with different number of elements (in case a friend was removed or added between old and new list generation). Also considering that some friends be have been logged to their account even if the first friend of old and new list are the same. The algorithm now searches for the first index that differs between old and new list and start detecting from this index